requirements.txt
appdirs==1.4.4
bandit==1.7.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.14
googlemaps==4.4.5
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pbr==5.5.1
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
smmap==4.0.0
stevedore==3.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
